Extract S3StateStoreTestBase
