+++ java/statestore/src/test/java/sleeper/statestore/s3/S3StateStoreTestBase.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2022-2023 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sleeper.statestore.s3;
+
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.io.TempDir;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+import sleeper.configuration.properties.instance.InstanceProperties;
+import sleeper.dynamodb.tools.DynamoDBContainer;
+
+import java.nio.file.Path;
+
+import static sleeper.configuration.properties.InstancePropertiesTestHelper.createTestInstanceProperties;
+import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.DATA_BUCKET;
+import static sleeper.configuration.properties.instance.CommonProperty.FILE_SYSTEM;
+import static sleeper.configuration.properties.instance.CommonProperty.MAXIMUM_CONNECTIONS_TO_S3;
+import static sleeper.dynamodb.tools.GenericContainerAwsV1ClientHelper.buildAwsV1Client;
+
+@Testcontainers
+public abstract class S3StateStoreTestBase {
+    @Container
+    public static DynamoDBContainer dynamoDb = new DynamoDBContainer();
+    protected static AmazonDynamoDB dynamoDBClient;
+    protected final InstanceProperties instanceProperties = createTestInstanceProperties();
+
+    @TempDir
+    public Path tempDir;
+
+    @BeforeAll
+    public static void initDynamoClient() {
+        dynamoDBClient = buildAwsV1Client(dynamoDb, dynamoDb.getDynamoPort(), AmazonDynamoDBClientBuilder.standard());
+    }
+
+    @AfterAll
+    public static void shutdownDynamoClient() {
+        dynamoDBClient.shutdown();
+    }
+
+    @BeforeEach
+    void setUp() {
+        instanceProperties.set(FILE_SYSTEM, "file://");
+        instanceProperties.setNumber(MAXIMUM_CONNECTIONS_TO_S3, 5);
+        instanceProperties.set(DATA_BUCKET, tempDir.toString());
+        new S3StateStoreCreator(instanceProperties, dynamoDBClient).create();
+    }
+}
